Updated container initialization
